api users register login logout

+++ app/Http/Controllers/ApiAuthController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\LoginUserRequest;
 use App\Http\Requests\LogoutUserRequest;
 use App\Http\Requests\RegisterUserRequest;
+use App\Http\Resources\UserResource;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
 
@@ -14,15 +15,29 @@ class ApiAuthController extends Controller
     {
         $validated = $request->validated();
 
+        if ($request->hasFile('avatar')) {
+            $validated['avatar'] = $request->file('avatar')->store('avatars', 'public');
+        }
+
+        if ($request->hasFile('id_card')) {
+            $validated['id_card'] = $request->file('id_card')->store('id_cards', 'public');
+        }
+
+        $validated['password'] = Hash::make($validated['password']);
+
         $user = User::create($validated);
-        // create user api token
-        $token = $user->createToken
-        ("api_token")->plainTextToken;
-        return response()->json([
-            'user' => $user,
+
+        $token = $user->createToken('api_token')->plainTextToken;
+
+        return UserResource::make($user)->additional([
+            'status' => 201,
+            'message' => 'User registered successfully',
+            'data' => [
             'token' => $token
-        ]);
+            ]
+        ])->response()->setStatusCode(201);
     }
+
     public function login(LoginUserRequest $request)
     {
         $validated = $request->validated();
@@ -33,30 +48,34 @@ class ApiAuthController extends Controller
             return response()->json([
                 'status' => 401,
                 'message' => 'Invalid credentials',
+                'data' => null
             ], 401);
         }
 
         $token = $user->createToken('api_token')->plainTextToken;
 
-        return response()->json([
+        return (new UserResource($user))->additional([
             'status' => 200,
-            'message' => 'Logged in',
+            'message' => 'Login ok',
             'data' => [
-                'user' => $user,
                 'token' => $token
             ]
         ]);
     }
     public function logout(LogoutUserRequest $request)
     {
-        $validated = $request->validated();
+        if (!$request->user()) {
+            return UserResource::make(null)->additional([
+                'status' => 401,
+                'message' => 'Unauthenticated',
+                'data' => null,
+            ])->response()->setStatusCode(401);
+        }
         $request->user()->currentAccessToken()->delete();
-        return response()->json(
-            [
-                'status' => 200,
-                'message' => 'Logged in',
-                'data' => null
-            ]
-        );
+        return UserResource::make(null)->additional([
+            'status' => 200,
+            'message' => 'Logged out',
+            'data' => null
+        ])->response()->setStatusCode(200);
     }
 }
